+++ CountlyRemoteConfigInternal.m
@@ -193,9 +193,14 @@ CLYRequestResult const CLYResponseError         = @"CLYResponseError";
         CLY_LOG_D(@"'fetchRemoteConfigForKeys' is aborted: SDK Networking is disabled from server config!");
         return;
     }
+    if (CountlyDeviceInfo.sharedInstance.isDeviceIDTemporary)
+    {
+        CLY_LOG_W(@"'fetchRemoteConfigForKeys' is aborted: device ID is in temporary mode!");
+        return;
+    }
     if (!completionHandler)
         return;
-    
+
     NSURLRequest* request = [self remoteConfigRequestForKeys:keys omitKeys:omitKeys isLegacy:isLegacy];
     NSURLSessionTask* task = [CountlyCommon.sharedInstance.ImmediateURLSession dataTaskWithRequest:request completionHandler:^(NSData* data, NSURLResponse* response, NSError* error)
                               {
@@ -493,9 +498,14 @@ CLYRequestResult const CLYResponseError         = @"CLYResponseError";
         CLY_LOG_D(@"'fetchVariantForKeys' is aborted: SDK Networking is disabled from server config!");
         return;
     }
+    if (CountlyDeviceInfo.sharedInstance.isDeviceIDTemporary)
+    {
+        CLY_LOG_W(@"'fetchVariantForKeys' is aborted: device ID is in temporary mode!");
+        return;
+    }
     if (!completionHandler)
         return;
-    
+
     NSURLRequest* request = [self downloadVariantsRequest];
     NSURLSessionTask* task = [CountlyCommon.sharedInstance.ImmediateURLSession dataTaskWithRequest:request completionHandler:^(NSData* data, NSURLResponse* response, NSError* error)
                               {
@@ -607,7 +617,13 @@ CLYRequestResult const CLYResponseError         = @"CLYResponseError";
         CLY_LOG_D(@"'enrollInRCVariant' is aborted: 'variantName' is not valid");
         return;
     }
-    
+
+    if (CountlyDeviceInfo.sharedInstance.isDeviceIDTemporary)
+    {
+        CLY_LOG_W(@"'enrollInRCVariant' is aborted: device ID is in temporary mode!");
+        return;
+    }
+
     NSURLRequest* request = [self enrollInVarianRequestForKey:key variantName:variantName];
     NSURLSessionTask* task = [CountlyCommon.sharedInstance.ImmediateURLSession dataTaskWithRequest:request completionHandler:^(NSData* data, NSURLResponse* response, NSError* error)
     {
@@ -720,9 +736,14 @@ CLYRequestResult const CLYResponseError         = @"CLYResponseError";
         CLY_LOG_D(@"'testingDownloaExperimentInfoInternal' is aborted: SDK Networking is disabled from server config!");
         return;
     }
+    if (CountlyDeviceInfo.sharedInstance.isDeviceIDTemporary)
+    {
+        CLY_LOG_W(@"'testingDownloaExperimentInfoInternal' is aborted: device ID is in temporary mode!");
+        return;
+    }
     if (!completionHandler)
         return;
-    
+
     NSURLRequest* request = [self downloadExperimentInfoRequest];
     NSURLSessionTask* task = [CountlyCommon.sharedInstance.ImmediateURLSession dataTaskWithRequest:request completionHandler:^(NSData* data, NSURLResponse* response, NSError* error)
                               {